CORRECTION2

+++ wallet/views.py
@@ -357,7 +357,8 @@ def activate(request, uidb64, token):
             invest(user)
             return redirect('/login')
         else:
-            user.delete()
+            if user.is_active is False:
+                user.delete()
             return render(request, 'wallet/home.html', {'message': 'Your Email was invalid, Therefore Your Account Has '
                                                                    'been deleted', 'status': 'danger'})
     except Investor.DoesNotExist:
